Eclipse code cleanup
(*) Mainly add @Override notation
(*) Remove some unnecessary casts
(*) Replace derived classes with base classes where appropriate

+++ 2.7.x/choicemaker-utils/choicemaker-util/src/main/java/com/choicemaker/util/IntValuedHashMap.java
@@ -42,7 +42,7 @@ public class IntValuedHashMap<K> extends HashMap<K, Integer> {
 	}
 
 	public int getInt(K key) {
-		Integer value = (Integer) get(key);
+		Integer value = get(key);
 		if (value != null)
 			return value.intValue();
 		else
